pyproject.toml
[project]
name = "comparison-tools"
version = "1.0.0"
description = "Smart comparison tools for apartments, homes, and hotels"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django~=6.0",
    "django-browser-reload~=1.21",
    "django-environ~=0.12",
    "gunicorn~=23.0",
    "social-auth-app-django~=5.6",
    "social-auth-core~=4.8",
    "google-cloud-secret-manager~=2.25",
    "google-cloud-logging~=3.11",
    "google-cloud-trace~=1.11",
    "google-cloud-profiler~=4.1",
    "google-auth~=2.43",
    "psycopg2-binary~=2.9",
    "stripe~=14.0",
    "geopy~=2.4",
    "googlemaps~=4.10",
]

[dependency-groups]
dev = [
    "ruff~=0.14",
]

[tool.ruff]
line-length = 120
target-version = "py312"
exclude = [
    ".git",
    ".venv",
    "venv",
    "__pycache__",
    "*/migrations/*",
    "theme/",
    "static/",
    "staticfiles/",
]

[tool.ruff.lint]
select = [
    "E",      # pycodestyle errors
    "W",      # pycodestyle warnings
    "F",      # Pyflakes
    "I",      # isort
    "B",      # flake8-bugbear
    "C4",     # flake8-comprehensions
    "UP",     # pyupgrade
    "DJ",     # flake8-django
]
ignore = [
    "E501",   # line too long (handled by formatter)
    "B008",   # do not perform function calls in argument defaults
    "DJ001",  # avoid using null=True on string-based fields
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py
"config/settings.py" = ["F401", "F403", "F405"]  # settings wildcards
"main.py" = ["E402"]  # import after Django setup is intentional for App Engine

[tool.ruff.lint.isort]
known-first-party = ["apartments", "config"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
